requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
docutils==0.20.1
frozenlist==1.4.1
geoip2==4.8.0
idna==3.10
influxdb==5.3.2
loguru==0.7.2
maxminddb==2.6.2
msgpack==1.1.0
multidict==6.1.0
propcache==0.2.0
pygtail==0.14.0
python-dateutil==2.9.0.post0
python-geohash==0.8.5
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
